requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Caching==2.3.0
redis==5.0.0
psycopg2-binary==2.9.7
cassandra-driver==3.25.0
flask-cors==3.0.10
pytest-flask==1.3.0